Refactor to common module
