+++ cloud/aws/cli/plugins/cloudtrail.py
@@ -1,13 +1,13 @@
 from invoke import task
 import dynaconf
 import core
+from common import COMMON_VALIDATORS, auto_app_fmt, conf, TFDIR
 
 
 VALIDATORS = [
-    dynaconf.Validator("VAST_AWS_REGION", must_exist=True),
+    *COMMON_VALIDATORS,
     dynaconf.Validator("VAST_CLOUDTRAIL_BUCKET_NAME", must_exist=True),
     dynaconf.Validator("VAST_CLOUDTRAIL_BUCKET_REGION", must_exist=True),
-    dynaconf.Validator("TF_STATE_BACKEND", default="local"),
 ]
 
 INVOKE_CONFIG = {}
@@ -18,8 +18,8 @@ def deploy(c, auto_approve=False):
     """Deploy the cloudtrail datasource"""
     core.init_step(c, "cloudtrail")
     c.run(
-        f"terragrunt apply {core.auto_app_fmt(auto_approve)} --terragrunt-working-dir {core.TFDIR}/cloudtrail",
-        env=core.conf(VALIDATORS),
+        f"terragrunt apply {auto_app_fmt(auto_approve)} --terragrunt-working-dir {TFDIR}/cloudtrail",
+        env=conf(VALIDATORS),
         pty=True,
     )
 
@@ -29,7 +29,7 @@ def destroy(c, auto_approve=False):
     """Remove the cloudtrail datasource"""
     core.init_step(c, "cloudtrail")
     c.run(
-        f"terragrunt destroy {core.auto_app_fmt(auto_approve)} --terragrunt-working-dir {core.TFDIR}/cloudtrail",
-        env=core.conf(VALIDATORS),
+        f"terragrunt destroy {auto_app_fmt(auto_approve)} --terragrunt-working-dir {TFDIR}/cloudtrail",
+        env=conf(VALIDATORS),
         pty=True,
     )
